Improvements to Python Tests (#879)
* Added conftest.py

Shared fixtures and helpers have been moved out of the monolithic test file into conftest.py so pytest can discover them across future test modules

* Broke up test_ingest_with_stats_v3 into focused tests

The ~400-line monolithic test covered 8 distinct behaviours across two APIs (read_variant_stats and read_allele_count). The test is now split into 15 tests and the repeated skipif condition was encapsulated into a shared module-level _skip_if_no_bcftools marker. Lastly, the equivalent read_variant_stats, read_variant_stats_arrow, and read_allele_count error checks were consolidated into loops for brevity.

* Replace manual use_arrow loops with pytest.mark.parametrize

Replacing the loops gives Pandas and Arrow their own named test paths, meaning failures are now reported distinctly.

* Split monolithic test file into focused modules

Additionally, _skip_if_no_bcftools was renamed to skip_if_no_bcftools and moved to conftest.py alongside the other shared helpers.

* Add tests to cover gaps in testing of create_dataset() parameters

The previously untested parameters include extra_attrs, tile_capacity, anchor_gap, allow_duplicates=False, checksum_type="md5", and variant_stats_version=2. Also, error-path tests were added for invalid checksum_type, supplying both extra_attrs and vcf_attrs, and calling create_dataset() on an already-existing dataset.

* Add tests to cover gaps in testing of ingest_samples() parameters

The previously untested parameters and execution paths include sample_uris=None early return, scratch_space_path/scratch_space_size must be provided together, invalid contig_mode, contigs_to_keep_separate/contigs_to_allow_merging must be lists, resume=True, sample_batch_size, and memory/thread tuning parameters.

* Add tests to cover gaps in testing of ReadConfig parameters

The previously untested parameters include sort_regions, buffer_percentage, tiledb_tile_cache_percentage, and tiledb_config supplied as both a dict and a tiledb.Config object.

* Add tests to cover gaps in testing of conttinue_read() and continue_read_arrow() parameters

Specifically, the release_buffers=False case is now tested for both methods.

* Add tests for delete_samples() edge cases and consolidate delete tests

The previously untested delete_samples() paths include empty list is a noop, None raises TypeError, and a non-existent sample name raises RuntimeError.

* Add tests to cover gaps in testing of the static Dataset.delete() method

The previously untested Dataset.delete() parameters and error paths include testing that the config parameter is accepted and does not prevent deletion of a local dataset and that deleting a non-existent URI raises TileDBError.

* Add test for Dataset constructor invalid mode

Tests that constructing a Dataset with an unrecognised mode string raises an exception with a descriptive message.

* Add tests for config_logging()

Test all six valid log levels via parametrization, that an invalid level raises with a descriptive message, and that the log_file parameter is accepted.

* Split Dataset-level tests into test_dataset.py and added version/schema_version tests

* Add tests for Dataset.sample_count()

Test that sample_count() returns the correct value for v3 and v4 datasets and is consistent with len(samples()), and that calling it on a write-mode dataset raises.

* Add tests for Dataset.tiledb_stats()

Test that tiledb_stats() returns a non-empty valid JSON string in both read and write mode. Add a skipped test documenting a bug where the stats-not-enabled guard never fires because get_tiledb_stats_enabled is referenced without () making it always truthy.

* Add tests for Dataset.export()

Test all previously uncovered export() parameters including samples, regions, output_format, merge, output_path, samples_file, bed_file, and skip_check_samples. It also checks that exporting in write-mode raises and tests the merge=True without output_path error path.

* Add tests for deprecated TileDBVCFDataset class

Test that constructing TileDBVCFDataset emits a DeprecationWarning with the expected message, and that the instance remains fully functional as a Dataset after construction.

* Fix DeprecationWarnings from deprecated region parameter in tests

Replaced incidental uses of the deprecated positional "region" parameter with "regions=[...]" and wrapped intentional uses in pytest.warns().

* Test that delete_samples() raises when the dataset is open in read mode

* Test that total_memory_percentage is accepted by ingest_samples()

* Test read_iter() with samples_file and bed_file parameters

* Test sample_qc() samples and config parameters

* Test error paths for read_allele_frequency()

Specifically, invalid region format and empty region are now tested.

* Updated tests to ensure read/read_arrow/read_iter parameter parity

Specifically, the samples, samples_file, bed_file, set_af_filter, and scan_all_samples parameters are now all exercised across all three methods.

* Test attributes() invalid attr_type raises

* Test that set_af_filter and scan_all_samples parameters can be combined in read() and read_arrow()

* Add missing Python test docstrings and standardize their style across all tests

* Removed extraneous comments from Python tests

* Removed the TileDB TBB threads test

The test was always marked to always be skipped and TBB threads has been long deprecated in TileDB.
